Move event storage classes and interfaces under the `event.storage` package
